feat: add functional on the goods page
